Remove travel_planner, static food_ordering; add dynamic food_ordering
